Remove panics in dockyards deployment controller
Currently the controller panics when a deployment reference can not be
found. This is very common when deployments are created and the panic is
a shortcut that should never have been commited.

This removes the panics and marks the deploment references ready
condition as false when deployment references are missing.