Add command line flag for metrics bind address
Currently the component will start the metrics server on the default
bind address. This causes issue when running multiple components locally
and they all try to use the same listening address.

This adds a command line flag to set the metrics bind address. The
metrics server is disabled by default.