Update dockyards workload webhook to validate template reference
Currently the template reference in a dockyards workload can be changed
as the field itself is not immutable. This is by design as it is a valid
operation to create a workload without a template reference to allow
another reconciler to complete it.

This updates the dockyards workload webhook to ensure that an empty
reference can be updated but no other changes are considered valid.