Update workload template webhook to show less verbose errors
Currently when the workload template webhook fails to load or build
instances the full source of the template is shown as the invalid value.
This makes it very hard to see the actual error message in the output.

This updates the workload template webhook to only show the lines of the
errors whe loading or building an instance fails.