Update dockyards workload controller to support list of byte types
Currently the dockyards workload controller will pass through fields of
byte types without modification. This results in the set operations
failing because byte is not a valid json type.

This updates the dockyards workload controller to convert any list of
bytes to a base64 encoded string. The method that ensures that values
are valid in json gets renamed to indicate that it no longer just
ensures integers are valid.

DY-388 #done.