Update dockyards workload controller to reconcile ready condition
This adds a reconciling condition on workloads while any owned
kustomization or helm release resources are reconciling. Following the
pattern used by fluxcd the reconciling condition is deleted if any
owned resources are not reconciling. This means that it is enough to
simple check for the existance of a reconciling condition.

This updates the dockyards workload controller to reconcile the
ready condition. The ready condition on workloads is a summary of the
workload template reconciled and reconciling conditions.

DY-394 #done.