Fix intervals in cloud provider for cloud director workload template
Currently the cloud provider for cloud director workload template uses
an interval of 5 minutes. This interval to too long while the cluster is
being created and too short for drift detection.

This adds a separate retry interval to make the reconciliation try more
often on errors. This increases the default interval to not check for
drift as often.

DY-423 #done.