Update dockyards workload controller to wait for ready condition
Currently the dockyards workload controller will reconcile all workloads
for a cluster before the cluster is ready. This should only be done for
cluster components which are a dependency for the cluster to be ready.

This updates the dockyards workload controller to only reconcile cluster
component workloads before the cluster is ready. For normal workloads
the workload template reconciliation will be ignored and the condition
will be set to false.