Add test helper for mocked custom resource definitions
Using the test environment helper from the controller runtime package
requires access to custom resource definitions for all dependencies.
There does not really exist a good solution for accessing the custom
resource definitions of the dependencies of this project.

This adds a package with mocked custom resource definitions for the
dockyards and flux custom resources. This ensures that the custom
resources exists in the test environment but will allow any fields to be
set. This is not an optimal solution but good enough to start with.